+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Function documentation comments in header section.
 - Add argument aliases.
+- Add basic support for otherwise unhandled positional arguments.
 
 ### Fixed